+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/NotificationDialogWithLongText.jsx
@@ -1,10 +1,6 @@
 import React, { useState } from 'react';
 import Button from 'terra-button';
-import NotificationDialog, { NotificationDialogVariants } from 'terra-notification-dialog';
-
-const clickConfirm = () => {
-  console.log('You clicked confirm'); // eslint-disable-line no-console
-};
+import NotificationDialog from 'terra-notification-dialog';
 
 const NotificationDialogWithLongText = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -99,23 +95,24 @@ const NotificationDialogWithLongText = () => {
 
   return (
     <>
-      <NotificationDialog
-        variant={NotificationDialogVariants.WARNING}
-        isOpen={isOpen}
-        title="Make sure that the title relates directly to the choices."
-        startMessage={message}
-        acceptAction={{
-          text: 'Confirm',
-          onClick: clickConfirm,
-        }}
-        rejectAction={{
-          text: 'Close',
-          onClick: handleCloseModal,
-        }}
-        buttonOrder="acceptFirst"
-        emphasizedAction="accept"
-      />
-      <Button text="Trigger NotificationDialog" onClick={handleOpenModal} />
+      {isOpen && (
+        <NotificationDialog
+          variant="hazard-medium"
+          dialogTitle="Use a Title That Relates Directly to the Actions - Long Titles Should Be Avoided When Possible"
+          startMessage={message}
+          acceptAction={{
+            text: 'Emphasized Accept Action',
+            onClick: handleCloseModal,
+          }}
+          rejectAction={{
+            text: 'Reject Action',
+            onClick: handleCloseModal,
+          }}
+          buttonOrder="acceptFirst"
+          emphasizedAction="accept"
+        />
+      )}
+      <Button text="Trigger Notification Dialog" onClick={handleOpenModal} />
     </>
   );
 };